+++ include/Player.h
@@ -11,13 +11,13 @@ class Player {
 public:
     Player();
     
-    void init(int startX, int startY);
+    void init(uint8_t startX, uint8_t startY);
     void reset();
     void update(Map& map);
     void setDirection(Direction dir);
     
-    int getX() const { return x; }
-    int getY() const { return y; }
+    uint8_t getX() const { return x; }
+    uint8_t getY() const { return y; }
     Direction getDirection() const { return direction; }
     
     bool isAlive() const { return alive; }
@@ -25,11 +25,11 @@ public:
     void revive();
 
 private:
-    int x, y;                    // Позиция в клетках
+    int8_t x, y;                 // Позиция в клетках
     Direction direction;         // Текущее направление
     Direction nextDirection;     // Следующее направление (буферизация)
     bool alive;                  // Жив ли игрок
-    int moveCounter;             // Счетчик для управления скоростью
+    uint8_t moveCounter;         // Счетчик для управления скоростью
     
     bool canMove(Map& map, Direction dir);
 };
